fix (#285): Fixed upload bug in save state upload button
