+++ solardoc/frontend/src/components/editor/dropdown/EditorSandwichDropdown.vue
@@ -5,24 +5,15 @@ import SandwichMenuDarkModeSVG from '@/components/icons/SandwichMenuDarkModeSVG.
 import { useDarkModeStore } from '@/stores/dark-mode'
 import { useOverlayStateStore } from '@/stores/overlay-state'
 import { useCurrentFileStore } from '@/stores/current-file'
-import { useCurrentUserStore } from '@/stores/current-user'
-import { showInfoNotifFromObj } from '@/scripts/show-notif'
-import { ensureLoggedIn } from '@/scripts/ensure-logged-in'
-import { interceptErrors } from '@/errors/handler/error-handler'
-import { useLoadingStore } from '@/stores/loading'
-import constants from '@/plugins/constants'
 import { useRouter } from 'vue-router'
 import { ref } from 'vue'
-import { createEditorRemoteFileConnection } from '@/scripts/editor/file'
 import { handleCopy } from '@/scripts/handle-copy'
 import { openInNewWindow } from '@/router/open'
-import * as phoenixBackend from '@/services/phoenix/api-service'
+import { uploadOrSaveFile } from '@/scripts/editor/upload'
 
 const darkModeStore = useDarkModeStore()
-const currentUserStore = useCurrentUserStore()
 const currentFileStore = useCurrentFileStore()
 const overlayStateStore = useOverlayStateStore()
-const loadingStore = useLoadingStore()
 
 const dropdown = ref(null)
 const $router = useRouter()
@@ -42,47 +33,7 @@ function handleNewFileButtonClick() {
 
 async function handleSaveButtonClick() {
   closeDropdown()
-
-  // Indicates whether this is a new file or an existing file
-  const wasAlreadyUploaded = !!currentFileStore.fileId
-
-  loadingStore.lockLoading()
-  loadingStore.pushMsg(
-    wasAlreadyUploaded
-      ? constants.loadingMessages.savingFileName
-      : constants.loadingMessages.uploadingFile,
-  )
-  try {
-    await interceptErrors(
-      ensureLoggedIn($router).then(
-        async () => await currentFileStore.storeOnServer(currentUserStore.bearer!),
-      ),
-    )
-    if (wasAlreadyUploaded) {
-      showInfoNotifFromObj(constants.notifMessages.fileSaved)
-      loadingStore.popMsg(constants.loadingMessages.savingFileName)
-    } else {
-      // Set the path to the editor for remote files (won't cause a real reload but simply change the path)
-      await $router.push({
-        name: 'remote-editor',
-        params: { fileId: currentFileStore.fileId },
-      })
-
-      showInfoNotifFromObj(constants.notifMessages.fileUploaded)
-      loadingStore.popMsg(constants.loadingMessages.uploadingFile)
-      loadingStore.pushMsg(constants.loadingMessages.loadingEditor)
-
-      // Ensure the phoenix backend is reachable and create a connection to the remote SDS server
-      await phoenixBackend.ensurePhoenixBackendIsReachable()
-      await createEditorRemoteFileConnection()
-
-      loadingStore.popMsg(constants.loadingMessages.loadingEditor)
-    }
-  } catch (e) {
-    loadingStore.unlockLoading()
-    throw e
-  }
-  loadingStore.unlockLoading()
+  return uploadOrSaveFile($router)
 }
 
 let copyButtonTimeout: null | ReturnType<typeof setTimeout> = null
