+++ solardoc/frontend/src/services/render/gen/backend-rest-service.ts
@@ -6,6 +6,7 @@
  */
 import * as Oazapfts from 'oazapfts/lib/runtime'
 import * as QS from 'oazapfts/lib/runtime/query'
+
 export const defaults: Oazapfts.RequestOpts = {
   baseUrl: '/',
 }
